+++ README.md
@@ -2,7 +2,7 @@
 https://www.pyimagesearch.com/2018/08/13/opencv-people-counter/
 
 # Run a sample video
-python test.py --prototxt mobilenet_ssd/MobileNetSSD_deploy.prototxt --model mobilenet_ssd/MobileNetSSD_deploy.caffemodel --input videos/walking.mp4 --output output/output_03.avi
+python track.py --prototxt mobilenet_ssd/MobileNetSSD_deploy.prototxt --model mobilenet_ssd/MobileNetSSD_deploy.caffemodel --input videos/walking.mp4 --output output/output_03.avi
 
 ## Pipeline 
 Run CV human tracking to get the rectangle/ centroid
